pyproject.toml
[tool.poetry]
name = "plato"
version = "2023.5.2"
description = "Templating as a microservice"
authors = [
  "Tiago Santos",
  "Joana Teixeira <joana.teixeira@morphotech.com>",
  "Rita Mariquitos <rita.mariquitos@morphotech.com>",
  "Fábio André Damas <fabio.damas@morphotech.com>"
]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.13"

fastapi = { extras = ["standard"], version = "^0.115.13" }
pydantic-settings = "^2.10.0"
uvicorn = "^0.35.0"
psycopg2-binary = "^2.9.10"
alembic = "^1.16.2"
jsonschema = "^4.24.0"

# S3
smart-open = { version = "6.3.0", extras = ["aws"] }

# Database
sqlalchemy = { version = "^2.0", extras = ["postgresql_psycopg2binary"] }

#HTTP
accept-types = "0.4.1"

# Templating
WeasyPrint = "52"
qrcode = { version = "6.1", extras = ["pil"] }
babel = "^2.9.1"
num2words = "0.5.10"

[tool.poetry.group.dev.dependencies]
pytest = "~8.4.1"
coverage = "~7.4.4"
moto = { version = "5.1.8", extras = ["s3"] }
testcontainers = "^4.10.0"
blinker = "1.4"
pypdf = "^5.8.0"

[build-system]
requires = ["poetry>=1.2"]
build-backend = "poetry.masonry.api"
